+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Added support for `.ptb` and `.canary` links to [yuyo.links][].
+
 ### Fixed
 - [ModalOptions][yuyo.modals.ModalOptions] attributes now correct expose the value passed to the
   modal rather an internal descriptor.